Make compiling containers against their non-native mode an error.
Concretely, leaky containers now complain if they're compiled in
non-leaky mode.

Similarly, non-leaky ones complain if they're compiled in leaky
mode.
