+++ mycpp/gc_heap.cc
@@ -3,10 +3,10 @@
 #include <sys/mman.h>  // mprotect()
 
 #include "mycpp/gc_types.h"
-
 using gc_heap::Heap;
 using gc_heap::Local;
 using gc_heap::Obj;
+
 
 GLOBAL_STR(kEmptyString, "");
