+++ mycpp/switchy_containers.cc
@@ -1,0 +1,7 @@
+
+#if LEAKY_BINDINGS
+#include "mycpp/mylib_old.cc"
+#else
+#include "mycpp/gc_heap.cc"
+#endif
+
